Refactor using primary metrics and mediator
